refactor: playground use monorepo packages
