+++ playground/backend-integration/package.json
@@ -21,6 +21,7 @@
     "http-proxy-middleware": "^2.0.6",
     "typescript": "^5.0.4",
     "vite": "^4.3.0",
-    "vite-plugin-checker": "workspace:*"
+    "vite-plugin-checker": "workspace:*",
+    "@vite-plugin-checker/typescript": "workspace:*"
   }
 }
