+++ playground/backend-integration/vite.config.js
@@ -1,6 +1,7 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as ts } from '@vite-plugin-checker/typescript'
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -11,5 +12,5 @@ export default defineConfig({
       input: './src/main.tsx',
     },
   },
-  plugins: [react(), checker({ typescript: true })],
+  plugins: [react(), checker([ts()])],
 })
