+++ playground/config-default/package.json
@@ -21,6 +21,7 @@
     "eslint": "^8.11.0",
     "typescript": "^5.0.4",
     "vite": "^4.3.0",
-    "vite-plugin-checker": "workspace:*"
+    "vite-plugin-checker": "workspace:*",
+    "@vite-plugin-checker/eslint": "workspace:*"
   }
 }
