+++ playground/config-enableBuild-false/vite.config.js
@@ -1,16 +1,21 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   // config-edit-slot
   plugins: [
-    checker({
-      enableBuild: false,
-      eslint: {
-        lintCommand: 'eslint ./src --ext .ts',
-      },
-      // checker-edit-slot
-    }),
+    checker(
+      [
+        eslint({
+          lintCommand: 'eslint ./src --ext .ts',
+        }),
+        // checker-edit-slot
+      ],
+      {
+        enableBuild: false,
+      }
+    ),
   ],
 })
