+++ playground/config-initialIsOpen-false/vite.config.js
@@ -1,15 +1,20 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 export default defineConfig({
   plugins: [
-    checker({
-      overlay: {
-        initialIsOpen: false,
-      },
-      eslint: {
-        lintCommand: 'eslint ./src --ext .ts',
-      },
-    }),
+    checker(
+      [
+        eslint({
+          lintCommand: 'eslint ./src --ext .ts',
+        }),
+      ],
+      {
+        overlay: {
+          initialIsOpen: false,
+        },
+      }
+    ),
   ],
 })
