+++ playground/config-no-runtime-in-build/vite.config.js
@@ -1,12 +1,13 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 export default defineConfig({
   plugins: [
-    checker({
-      eslint: {
+    checker([
+      eslint({
         lintCommand: 'eslint ./src --ext .ts',
-      },
-    }),
+      }),
+    ]),
   ],
 })
