+++ playground/config-overlay-changes/vite.config.js
@@ -1,15 +1,16 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   // config-edit-slot
   plugins: [
-    checker({
-      eslint: {
+    checker([
+      eslint({
         lintCommand: 'eslint ./src --ext .ts',
-      },
+      }),
       // checker-edit-slot
-    }),
+    ]),
   ],
 })
