+++ playground/config-overlay-false/vite.config.js
@@ -1,14 +1,18 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
-
+import { checker } from 'vite-plugin-checker'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 export default defineConfig({
   plugins: [
-    checker({
-      overlay: false,
-      eslint: {
-        lintCommand: 'eslint ./src --ext .ts',
-      },
-    }),
+    checker(
+      [
+        eslint({
+          lintCommand: 'eslint ./src --ext .ts',
+        }),
+      ],
+      {
+        overlay: false,
+      }
+    ),
   ],
 })
