+++ playground/eslint-config-log-level/package.json
@@ -21,6 +21,7 @@
     "@typescript-eslint/parser": "^5.59.0",
     "typescript": "^5.0.4",
     "vite": "^4.3.0",
-    "vite-plugin-checker": "workspace:*"
+    "vite-plugin-checker": "workspace:*",
+    "@vite-plugin-checker/eslint": "workspace:*"
   }
 }
