+++ playground/multiple-reload/vite.config.ts
@@ -1,14 +1,16 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as ts } from '@vite-plugin-checker/typescript'
+import { checker as eslint } from '@vite-plugin-checker/eslint'
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [
-    checker({
-      typescript: true,
-      eslint: {
+    checker([
+      ts(),
+      eslint({
         lintCommand: 'eslint ./src --ext .ts',
-      },
-    }),
+      }),
+    ]),
   ],
 })
