+++ playground/stylelint-default/vite.config.js
@@ -1,12 +1,13 @@
 import { defineConfig } from 'vite'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as stylelint } from '@vite-plugin-checker/stylelint'
 
 export default defineConfig({
   plugins: [
-    checker({
-      stylelint: {
+    checker([
+      stylelint({
         lintCommand: 'stylelint "./**/*.css"',
-      },
-    }),
+      }),
+    ]),
   ],
 })
