+++ playground/vue-tsc-vue3-ts4/vite.config.js
@@ -1,8 +1,9 @@
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
-import checker from 'vite-plugin-checker'
+import { checker } from 'vite-plugin-checker'
+import { checker as vueTsc } from '@vite-plugin-checker/vue-tsc'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), checker({ vueTsc: true })],
+  plugins: [vue(), checker([vueTsc()])],
 })
